refactor: notion 패키지 calendar로 이동하고 properties 바인딩 적용
