+++ backend/src/main/java/moadong/calendar/notion/payload/response/NotionTokenExchangeResponse.java
@@ -1,4 +1,4 @@
-package moadong.integration.notion.payload.response;
+package moadong.calendar.notion.payload.response;
 
 public record NotionTokenExchangeResponse(
         String workspaceName,
