+++ backend/src/main/java/moadong/calendar/notion/repository/NotionConnectionRepository.java
@@ -1,6 +1,6 @@
-package moadong.integration.notion.repository;
+package moadong.calendar.notion.repository;
 
-import moadong.integration.notion.entity.NotionConnection;
+import moadong.calendar.notion.entity.NotionConnection;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface NotionConnectionRepository extends MongoRepository<NotionConnection, String> {
